vespa_util/pyproject.toml
[tool.poetry]
name = "vespa-util"
version = "0.1.0"
description = ""
authors = ["Jakub 'lambdaofgod' Bartczuk <bartczukkuba@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyvespa = "^0.39.0"
langchain = "^0.1.8"
pypdf = "^4.0.2"
openai = "^1.12.0"
fire = "^0.5.0"
pydantic = "^2.6.1"
tqdm = "^4.66.2"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
